requirements.txt
tqdm
pytorch
pytorch_lightning
getpaths
matplotlib